WINDUP-1145 - Added refresh feature to issue explorer for re-generating issue markers, and re-populating the issue explorer. Added timestamp to run configuration to support removing issue markers from files when the most recent configuration is deleted.

+++ plugins/org.jboss.tools.windup.ui/src/org/jboss/tools/windup/ui/internal/services/LaunchConfigurationService.java
@@ -59,7 +59,10 @@ public class LaunchConfigurationService implements ILaunchConfigurationListener 
 	@Override
 	public void launchConfigurationRemoved(ILaunchConfiguration launchConfig) {
 		ConfigurationElement configuration = modelService.findConfiguration(launchConfig.getName());
-		markerService.deleteAllWindupMarkers();
+		ConfigurationElement lastConfiguration = modelService.getRecentConfiguration();
+		if (configuration.getName().equals(lastConfiguration.getName())) {
+			markerService.deleteAllWindupMarkers();
+		}
 		modelService.deleteConfiguration(configuration);
 	}
